requirements_final.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
asttokens==3.0.0
attrs==24.3.0
axial_positional_embedding==0.3.4
certifi==2024.12.14
charset-normalizer==3.4.1
CoLT5-attention==0.11.1
comet-ml==3.47.6
comm==0.2.2
configobj==5.0.9
debugpy==1.6.7
decorator==5.1.1
dnspython==2.7.0
dulwich==0.22.7
einops==0.8.0
everett==3.1.0
exceptiongroup==1.2.2
executing==2.1.0
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.12.0
grpcio==1.69.0
idna==3.10
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.31.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
lightning==2.5.0.post0
lightning-utilities==0.11.9
local-attention==1.9.15
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
nest_asyncio==1.6.0
networkx==3.4.2
numpy==2.2.1
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pip==24.2
platformdirs==4.3.6
product_key_memory==0.2.11
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==5.29.2
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.0
pymongo==4.10.1
python-box==6.1.0
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.0.post0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
reformer-pytorch==1.4.4
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
scikit-learn==1.6.0
scipy==1.15.0
semantic-version==2.10.0
sentry-sdk==2.19.2
setuptools==75.1.0
simplejson==3.19.3
six==1.17.0
stack_data==0.6.3
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
torch==2.5.1
torchmetrics==1.6.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
triton==3.1.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.44.0
wrapt==1.17.0
wurlitzer==3.1.1
yarl==1.18.3
zipp==3.21.0
